requirements.txt
fastapi
gymnasium
matplotlib
numpy==2.0.2
pandas
python-dotenv
stable_baselines3==2.6.0
ta
joblib
google-cloud-storage
google-cloud-secret-manager
protobuf==3.20.3
firebase_admin
feedparser
newspaper3k
lxml_html_clean
groq
pydantic
python-binance
redis
langfuse
yfinance